+++ index.php
@@ -14,7 +14,7 @@ if(isset($REQ["reset_email"])) {
     die();
 }
 if(isset($REQ["language"])) {
-    $_SESSION["language"] = $REQ["language"];
+    setcookie("language", $REQ["language"],false,"",$_SERVER['SERVER_NAME'],true,true);
     header("Location: /");
     exit;
 }
@@ -35,9 +35,9 @@ if(
         $login_type = "customer";
     }
     if($login->success) {
-        setcookie("ipp_type",  $login_type, time()+3600);  /* expire in 1 hour */
-        setcookie("ipp_id", $login->content->user_id, time()+3600);  /* expire in 1 hour */
-        setcookie("ipp_session_id", $login->content->session_id, time()+3600);  /* expire in 1 hour */
+        setcookie("ipp_type",  $login_type, time()+3600,"",$_SERVER['SERVER_NAME'],true,true);  /* expire in 1 hour */
+        setcookie("ipp_id", $login->content->user_id, time()+3600,"",$_SERVER['SERVER_NAME'],true,true);  /* expire in 1 hour */
+        setcookie("ipp_session_id", $login->content->session_id, time()+3600,"",$_SERVER['SERVER_NAME'],true,true);  /* expire in 1 hour */
         header("location: $url");
         exit;
     }